+++ crates/medea-macro/src/dart_bridge.rs
@@ -493,21 +493,15 @@ impl FnExpander {
     /// # Example of the generated code
     ///
     /// ```ignore
-    /// PEER_CONNECTION__CREATE_OFFER__FUNCTION
-    ///     .set(SendWrapper::new(create_offer))
-    ///     .expect(
-    ///         "PEER_CONNECTION__CREATE_OFFER__FUNCTION \
-    ///         can only be set once",
-    ///     );
+    /// *PEER_CONNECTION__CREATE_OFFER__FUNCTION.borrow_mut() =
+    ///     Some(create_offer);
     /// ```
     fn gen_register_fn_expr(&self) -> syn::Expr {
         let fn_storage_ident = &self.fn_storage_ident;
         let ident = &self.ident;
-        let expect_message = format!("{fn_storage_ident} can only be set once");
 
         parse_quote! {
-            #fn_storage_ident.set(::send_wrapper::SendWrapper::new(#ident))
-                .expect(#expect_message)
+            *#fn_storage_ident.borrow_mut() = Some(#ident)
         }
     }
 
@@ -534,18 +528,23 @@ impl FnExpander {
     /// # Example of generated code
     ///
     /// ```ignore
-    /// static PEER_CONNECTION__CREATE_OFFER__FUNCTION: OnceLock<
-    ///     SendWrapper<PeerConnectionCreateOfferFunction>,
-    /// > = OnceLock::new();
+    /// static PEER_CONNECTION__CREATE_OFFER__FUNCTION: LazyLock<
+    ///     SendWrapper<RefCell<Option<PeerConnectionCreateOfferFunction>>>,
+    /// > = LazyLock::new(|| SendWrapper::new(RefCell::new(None)));
     /// ```
     fn gen_fn_storages(&self) -> TokenStream {
         let name = &self.fn_storage_ident;
         let type_alias = &self.type_alias_ident;
 
         quote! {
-            static #name: ::std::sync::OnceLock<
-                ::send_wrapper::SendWrapper<#type_alias>> =
-                    ::std::sync::OnceLock::new();
+            static #name: ::std::sync::LazyLock<
+                ::send_wrapper::SendWrapper<
+                    ::std::cell::RefCell<Option<#type_alias>>>> =
+                        ::std::sync::LazyLock::new(|| {
+                            ::send_wrapper::SendWrapper::new(
+                                ::std::cell::RefCell::new(None)
+                            )
+                        });
         }
     }
 
@@ -557,10 +556,10 @@ impl FnExpander {
     /// pub unsafe fn create_offer(
     ///     peer: Dart_Handle,
     /// ) -> Result<Dart_Handle, Error> {
-    ///     let res = (PEER_CONNECTION__CREATE_OFFER__FUNCTION
-    ///         .get()
+    ///     let res = (*PEER_CONNECTION__CREATE_OFFER__FUNCTION
+    ///         .borrow()
     ///         .as_ref()
-    ///         .expect("PEER_CONNECTION__CREATE_OFFER__FUNCTION is not set"))
+    ///         .expect("`PEER_CONNECTION__CREATE_OFFER__FUNCTION` is not set"))
     ///         (peer);
     ///     if let Some(e) = PEER_CONNECTION__CREATE_OFFER__ERROR
     ///         .borrow_mut()
@@ -589,20 +588,20 @@ impl FnExpander {
 
         let ret_ty = &self.ret_ty;
         let fn_storage_ident = &self.fn_storage_ident;
-        let expect_message = format!("`{fn_storage_ident}` is not set");
+        let none_message = format!("`{fn_storage_ident}` is not set");
 
         quote! {
             #( #doc_attrs )*
             pub unsafe fn #name(#args) #ret_ty {
-                let res = (
-                        #fn_storage_ident.get()
-                            .as_ref()
-                            .expect(#expect_message)
-                    )(#( #args_idents ),*);
+                let res = (*#fn_storage_ident
+                    .borrow()
+                    .as_ref()
+                    .expect(#none_message))
+                    (#( #args_idents ),*);
                 if let Some(e) = #error_slot.borrow_mut().take() {
-                  Err(e)
+                    Err(e)
                 } else {
-                  Ok(res)
+                    Ok(res)
                 }
             }
         }