+++ flutter/lib/src/web/jason.dart
@@ -18,8 +18,13 @@ Future<void> setLogLevel(base.LogLevel level) async {
 class Jason extends base.Jason {
   final wasm.Jason obj = wasm.Jason();
 
+  /// Initializes the plugin's FFI layer.
+  ///
+  /// No-op for Web.
+  static Future<void> ensureInitialized() async {}
+
   /// Creates a new instance of [Jason].
-  static Future<Jason> init() async {
+  static Jason create() {
     return Jason._();
   }