+++ flutter/CHANGELOG.md
@@ -8,15 +8,23 @@ All user visible changes to this project will be documented in this file. This p
 
 ## master
 
+### BC Breaks
+
+- `Jason.init()` decoupled into two separate methods: ([#267])
+    - `Jason.ensureInitialized()` - async idempotent FFI layer initialization.
+    - `Jason.create()` - sync static constructor.
+
 ### Added
 
 - `MediaDeviceDetails.sampleRate()` and `MediaDeviceDetails.numChannels()` methods (only works on Windows right now). ([#264])
 
 ### Fixed
 
+- Segfault on [Flutter] hot restart. ([#267])
 - `MediaDeviceDetails.groupId()` not working on Windows platform. ([#264])
 
 [#264]: https://github.com/instrumentisto/medea-jason/pull/264
+[#267]: https://github.com/instrumentisto/medea-jason/pull/267